KTOR-2763 Fix Warnings and Enable -Werror flag (#2496)
* KTOR-2763 Fix warnings in ktor-io

* KTOR-2763 Fix warnings in ktor-utils

* KTOR-2763 Fix warnings in ktor-http

* KTOR-2763 Fix warnings in ktor-network

* KTOR-2763 Fix warnings in ktor-features

* KTOR-2763 Fix warnings in ktor-server

* KTOR-2763 Fix warnings in ktor-client

* KTOR-2763 Fix warnings in ktor-shared & ktor-test-dispatcher

* KTOR-2763 Update configuration

+++ ktor-io/jvm/src/io/ktor/utils/io/internal/Strings.kt
@@ -114,7 +114,7 @@ private fun ByteBuffer.decodeASCII3_array(out: CharArray, offset: Int, length: I
             val b = array[srcPos]
             if (b < 0) break
 
-            out[pos] = b.toChar()
+            out[pos] = b.toInt().toChar()
 
             pos++
             srcPos++
@@ -143,7 +143,7 @@ private fun ByteBuffer.decodeASCII3_buffer(out: CharArray, offset: Int, length: 
                 pushBack = true
                 break
             }
-            out[pos] = b.toChar()
+            out[pos] = b.toInt().toChar()
             pos++
         }
     }
@@ -172,7 +172,7 @@ private inline fun ByteBuffer.decodeASCII3_array(
             val b = array[srcPos]
             if (b < 0) break
 
-            val ch = b.toChar()
+            val ch = b.toInt().toChar()
             if (!predicate(ch)) {
                 position(srcPos - arrayOffset())
                 return decodeUtf8Result(pos - offset, -1)
@@ -213,7 +213,7 @@ private inline fun ByteBuffer.decodeASCII3_buffer(
                 break
             }
 
-            val ch = b.toChar()
+            val ch = b.toInt().toChar()
             if (!predicate(ch)) {
                 pushBack = true
                 predicateFailed = true
